extconf: avoid unnecessary linkage against libkqueue
The check for libkqueue may succeed out-of-the-box on Debian systems
where libkqueue-dev is installed.  However, libkqueue-dev on Debian
installs sys/event.h in a non-standard include path
(/usr/include/kqueue), so it is not picked up by default and
kqueue support is never compiled.

So only check for (and link to) libkqueue if we are configured
to detect sys/event.h.  This should not affect users on *BSD
platforms with native kqueue support.